Factor `DataModel` into its own module
