+++ src/roman_datamodels/datamodels/__init__.py
@@ -1,1 +1,2 @@
+from ._core import *  # noqa: F403
 from ._datamodels import *  # noqa: F403
